django_project/requirements.txt
Django==2.2.*
Pillow==6.2.*
djangorestframework==3.10.*
gunicorn==20.0.4
pytz==2019.3
sqlparse==0.3.0
django-import-export==2.0
django-heroku==0.3.1

